sdk/pyproject.toml
[tool.poetry]
name = "pv-forecast-sdk"
version = "0.1.0"
description = ""
authors = ["ulrik christensen <ulrikchristensen@outlook.dk>"]
readme = "README.md"
packages = [
    { include = "sdk" },
]

[tool.poetry.dependencies]
python = "^3.10"
azure-identity = "^1.18.0"
xarray = {extras = ["complete"], version = "^2024.9.0"}
azure-storage-file-datalake = "^12.17.0"
polars = "^1.9.0"
duckdb = "^1.1.1"
scipy = "^1.14.1"
cftime = "^1.6.4"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
